додав продукту властивість AvailableAmount, змінив через це функції видалення продуктів, а також створення, видалення, відхилення та відправлення замовлення

+++ Infrastructure/Services/WarehouseCustomersService.cs
@@ -1,6 +1,7 @@
 ﻿using Infrastructure.Interfaces;
 using Models.Api.Common.Response;
 using Models.Api.Customer.Request;
+using Models.Api.Customer.Response;
 using Models.DBModels;
 using Models.DBModels.Enums;
 
@@ -10,37 +11,48 @@ namespace Infrastructure.Services
     {
         public WarehouseCustomersService(IProductsRepository productsRepository, IOrdersRepository ordersRepository, ISessionsRepository sessionsRepository) : base(productsRepository, ordersRepository, sessionsRepository) { }
 
-        public ActionWithOrderSuccessModel MakeOrder(CreateOrderRequestModel createRequest)
+        public CreateOrderResponseModel MakeOrder(CreateOrderRequestModel createRequest)
         {
-            Order newOrder = new Order()
+            Product orderedProduct = _productsRepository.GetProduct(createRequest.Product.ProductId);
+            Order newOrder = new Order
             {
                 Status = OrderStatus.Created,
-                Product = createRequest.Product,
+                Product = orderedProduct,
                 Quantity = createRequest.Quantity,
                 OrderPrice = createRequest.Product.Price * createRequest.Quantity,
                 User = createRequest.User
             };
             _ordersRepository.CreateOrder(newOrder);
+            orderedProduct.AvailableAmount -= (int)createRequest.Quantity;
+            _productsRepository.UpdateProduct(orderedProduct);
             var addedOrder = _ordersRepository.GetJustCreatedOrder(newOrder); // to get orderId
             return new() 
             {
-                Order = addedOrder
+                Order = addedOrder,
+                Success = true,
+                Message = "Order created successfully"
             };
         }
 
-        public ActionWithOrderSuccessModel RemoveOrder(RemoveOrderRequestModel removeOrderRequest)
+        public RemoveOrderResponseModel RemoveOrder(RemoveOrderRequestModel removeOrderRequest)
         {
             var deletedOrder = _ordersRepository.GetOrder(removeOrderRequest.OrderId);
+            if (deletedOrder.User.UserId != removeOrderRequest.UserId)
+                return new() { Success = false, Message = "can't remove another user's order", Order = deletedOrder };
+            if (deletedOrder.Status == OrderStatus.Sent)
+                return new() { Success = false, Message = "can't remove already sent order", Order = deletedOrder };
+            if (deletedOrder.Status == OrderStatus.Rejected)
+                return new() { Success = false, Message = "this order is already rejected", Order = deletedOrder };
+            Product orderedProduct = _productsRepository.GetProduct(deletedOrder.Product.ProductId.Value!);
+            orderedProduct.AvailableAmount += (int)deletedOrder.Quantity;
             _ordersRepository.DeleteOrder(deletedOrder);
+            _productsRepository.UpdateProduct(orderedProduct);
             return new()
             {
-                Order = deletedOrder
+                Order = deletedOrder,
+                Success = true,
+                Message = "Order deleted successfully"
             };
-        }
-
-        public string GetProductInfo()
-        {
-            throw new NotImplementedException();
         }
     }
 }
